modularization; tests
